[WIP] Large refactoring of ttbar analysis
The goal of the refactoring was to make the actual analysis logic stand
out more, making the code easier to follow.

The main changes are:

- class TtbarAnalysis has been removed:
  its logic has been split between a few free functions
- a few helper functions and types have been added to utils.py, so that
  analysis.py contains almost only the actual analysis logic
- plotting logic has been moved to plotting.py

The new implementation is broken with distributed RDF.
Otherwise it is expected to produce exactly the same
plots and output histograms as before the refactoring.