Use SetRangeUser instead of slicing histograms
This removes the dependency of plotting.py on the
jitted ROOT.Slice function.

Slicing and then rebinning is not strictly
equivalent to rebinning and then reducing the axis
range so this change introduces minor changes in
the aspect of the btag.png plot.
However the reference implementation does not
specify how the histograms should be plotted so I
think it is ok: the contents of the histograms
saved in output are not modified.